// angular/libs/cms-kit-management/config/src/enums/policy-names.ts
export const enum eCmsManagementPolicyNames {
  CmsKitManagement = '',
  Comments = 'CmsKit.Comments',
  Tags = 'CmsKit.Tags',
  Ratings = 'CmsKit.Ratings',
  Reactions = 'CmsKit.Reactions',
  Contents = 'CmsKit.Contents',
  Pages = 'CmsKit.Pages',
  Blogs = 'CmsKit.Blogs',
  BlogPosts = 'CmsKit.BlogPosts',
  MediaDescriptors = 'CmsKit.MediaDescriptors',
  Menus = 'CmsKit.Menus'
}